refactor(front): move job specific components
Move job specific components from jobs folder to job folder. The jobs
folder contains components used in jobs view only.
