+++ frontend/src/components/job/JobStatusLabel.vue
