+++ frontend/src/views/JobView.vue
@@ -14,8 +14,8 @@ import type { LocationQueryRaw } from 'vue-router'
 import ClusterMainLayout from '@/components/ClusterMainLayout.vue'
 import { useClusterDataPoller } from '@/composables/DataPoller'
 import type { ClusterIndividualJob } from '@/composables/GatewayAPI'
-import JobStatusLabel from '@/components/jobs/JobStatusLabel.vue'
-import JobProgress from '@/components/jobs/JobProgress.vue'
+import JobStatusLabel from '@/components/job/JobStatusLabel.vue'
+import JobProgress from '@/components/job/JobProgress.vue'
 import { useRuntimeStore } from '@/stores/runtime'
 import ErrorAlert from '@/components/ErrorAlert.vue'
 import LoadingSpinner from '@/components/LoadingSpinner.vue'
