+++ frontend/src/views/JobsView.vue
@@ -15,7 +15,7 @@ import type { JobSortCriterion, JobSortOrder } from '@/stores/runtime'
 import { useClusterDataPoller } from '@/composables/DataPoller'
 import type { ClusterJob } from '@/composables/GatewayAPI'
 import JobsSorter from '@/components/jobs/JobsSorter.vue'
-import JobStatusLabel from '@/components/jobs/JobStatusLabel.vue'
+import JobStatusLabel from '@/components/job/JobStatusLabel.vue'
 import ClusterMainLayout from '@/components/ClusterMainLayout.vue'
 import JobsFiltersPanel from '@/components/jobs/JobsFiltersPanel.vue'
 import JobsFiltersBar from '@/components/jobs/JobsFiltersBar.vue'
