+++ frontend/src/views/NodeView.vue
@@ -16,7 +16,7 @@ import type { ClusterDataPoller } from '@/composables/DataPoller'
 import type { ClusterIndividualNode, ClusterJob } from '@/composables/GatewayAPI'
 import NodeMainState from '@/components/resources/NodeMainState.vue'
 import NodeAllocationState from '@/components/resources/NodeAllocationState.vue'
-import JobStatusLabel from '@/components/jobs/JobStatusLabel.vue'
+import JobStatusLabel from '@/components/job/JobStatusLabel.vue'
 import ErrorAlert from '@/components/ErrorAlert.vue'
 import LoadingSpinner from '@/components/LoadingSpinner.vue'
 import { ChevronLeftIcon } from '@heroicons/vue/20/solid'
